requirements.txt
opencv-python>=4.7.0.72
PyQt6>=6.5.1
numpy>=1.24.3
torch>=2.0.1
Pillow>=9.5.0
ultralytics>=8.0.114